+++ cmd/backup.go
@@ -1,21 +1,18 @@
 package main
 
 import (
-	"bufio"
-	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
 	"os"
-	"os/exec"
 	"path"
 	"runtime"
-	"strings"
 
 	"github.com/ansel1/merry/v2"
-	"github.com/go-logr/logr"
+	"rclone-backup.kimi450.com/pkg/config"
 	"rclone-backup.kimi450.com/pkg/io"
 	"rclone-backup.kimi450.com/pkg/logging"
+	"rclone-backup.kimi450.com/pkg/rclone"
 )
 
 type ScriptArgs struct {
@@ -28,37 +25,7 @@ type ScriptArgs struct {
 	Checksum bool
 }
 
-type LogFile struct {
-	Path string
-	File *os.File
-}
-
-type BackupConfig struct {
-	Name      string `json:"name"`
-	SourceDir string `json:"sourceDir"`
-	DestDir   string `json:"destDir"`
-}
-
-type Config struct {
-	Items []BackupConfig `json:"items"`
-}
-
-func RunRcloneVersion(log logr.Logger, rcloneBinary, rcloneConfig string) error {
-	cmd := exec.Command(rcloneBinary, "version",
-		"--config", rcloneConfig)
-	log.Info("running command", "cmd", cmd.String())
-
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return merry.Errorf("failed to run command: %s: %w", string(output), err)
-	}
-	log.Info("finished running command", "cmd", cmd.String(),
-		"output", string(output))
-
-	return nil
-}
-
-func parseArgs() (*ScriptArgs, error) {
+func (scriptArgs *ScriptArgs) parseArgs() error {
 	flag.Usage = func() {
 		fmt.Fprintf(os.Stderr, "Usage of %s:\n", os.Args[0])
 
@@ -68,7 +35,7 @@ func parseArgs() (*ScriptArgs, error) {
 
 	cwd, err := os.Getwd()
 	if err != nil {
-		return nil, merry.Errorf("failed to get working directory: %w", err)
+		return merry.Errorf("failed to get working directory: %w", err)
 	}
 
 	// linux
@@ -90,7 +57,7 @@ func parseArgs() (*ScriptArgs, error) {
 		`Location of the clone config file
 		
 When using a remote source you are expected to have set it up already using '.\rclone.exe config'.
-This remote's name is to be used in the the -BackupConfigJson file as the SourceDir.
+This remote's name is to be used in the the config file as the SourceDir.
 Example: 'google-drive:'
 
 Refer to these pages to setup drive and the recommended
@@ -111,19 +78,22 @@ client-id and client-secret required for this setup:
 
 	flag.Parse()
 
-	args := &ScriptArgs{
-		LogBundleBaseDir: *logBundleBaseDir,
-		RcloneBinary:     *rcloneBinary,
-		RcloneConfig:     *rcloneConfig,
-		Config:           *config,
-		DryRun:           *dryRun,
-		Checksum:         *checksum,
-	}
-	return args, nil
+	scriptArgs.LogBundleBaseDir = *logBundleBaseDir
+	scriptArgs.RcloneBinary = *rcloneBinary
+	scriptArgs.RcloneConfig = *rcloneConfig
+	scriptArgs.Config = *config
+	scriptArgs.DryRun = *dryRun
+	scriptArgs.Checksum = *checksum
+
+	return nil
 }
 
-func validateArgs(args *ScriptArgs) error {
-	filePaths := []string{args.Config, args.RcloneBinary, args.RcloneConfig}
+func (scriptArgs *ScriptArgs) verifyExpectedFilesExist() error {
+	filePaths := []string{
+		scriptArgs.Config,
+		scriptArgs.RcloneBinary,
+		scriptArgs.RcloneConfig,
+	}
 
 	for _, filePath := range filePaths {
 		if _, err := os.Stat(filePath); errors.Is(err, os.ErrNotExist) {
@@ -134,175 +104,7 @@ func validateArgs(args *ScriptArgs) error {
 	return nil
 }
 
-func SyncSourceAndDestination(log logr.Logger, logBundleDir string,
-	rcloneBinary, rcloneConfig string,
-	extraSyncArgs []string, backupConfig BackupConfig,
-) error {
-	fileDateTime := io.GetDateTimePrefixForFile()
-
-	logFileSourceFiles := &LogFile{
-		Path: path.Join(logBundleDir,
-			fmt.Sprintf("%s-%s-source-files.json",
-				fileDateTime, backupConfig.Name)),
-	}
-	logFileDestFilesBeforeSync := &LogFile{
-		Path: path.Join(logBundleDir,
-			fmt.Sprintf("%s-%s-dest-files-before-sync.json",
-				fileDateTime, backupConfig.Name)),
-	}
-	logFileDestFilesAfterSync := &LogFile{
-		Path: path.Join(logBundleDir,
-			fmt.Sprintf("%s-%s-dest-files-after-sync.json",
-				fileDateTime, backupConfig.Name)),
-	}
-	logFileSync := &LogFile{
-		Path: path.Join(logBundleDir,
-			fmt.Sprintf("%s-%s-sync-logs.json",
-				fileDateTime, backupConfig.Name)),
-	}
-	logFileSyncCombinedReport := &LogFile{
-		Path: path.Join(logBundleDir,
-			fmt.Sprintf("%s-%s-sync-report.txt",
-				fileDateTime, backupConfig.Name)),
-	}
-
-	logFiles := []*LogFile{
-		logFileSourceFiles,
-		logFileDestFilesBeforeSync,
-		logFileDestFilesAfterSync,
-		logFileSync,
-		logFileSyncCombinedReport,
-	}
-
-	for _, logFile := range logFiles {
-		log.Info("creating log file", "filePath", logFile.Path)
-		file, err := os.Create(logFile.Path)
-		if err != nil {
-			return merry.Errorf("failed to create log file '%s': %s",
-				logFile.Path, err)
-		}
-		logFile.File = file
-		defer logFile.File.Close()
-	}
-
-	// Get all the files in source dir
-	log.Info("getting ls data for source directory", "path", backupConfig.SourceDir)
-	err := RunRcloneLsJson(log, logFileSourceFiles.File, rcloneBinary,
-		rcloneConfig, backupConfig.SourceDir)
-	if err != nil {
-		log.Info("WARN: directory does not exist before sync, skipped getting ls data for dir",
-			"path", backupConfig.SourceDir)
-	}
-
-	// Get all the files in destination dir before syncing data
-	log.Info("getting ls data for destination directory before syncing data",
-		"path", backupConfig.DestDir)
-	err = RunRcloneLsJson(log, logFileDestFilesBeforeSync.File, rcloneBinary,
-		rcloneConfig, backupConfig.DestDir)
-	if err != nil {
-		log.Info("WARN: directory does not exist before sync, skipped getting ls data for dir",
-			"path", backupConfig.DestDir)
-	}
-
-	// Run the sync command
-	log.Info("syncing source and destination",
-		"source", backupConfig.SourceDir,
-		"destination", backupConfig.DestDir,
-		"syncLogs", logFileSync.Path,
-		"syncReport", logFileSyncCombinedReport.Path,
-	)
-	err = RunRcloneSync(log, logFileSync.Path, logFileSyncCombinedReport.Path,
-		rcloneBinary, rcloneConfig, backupConfig.SourceDir, backupConfig.DestDir,
-		extraSyncArgs)
-	if err != nil {
-		return merry.Errorf("failed to sync source and destination: %w", err)
-	}
-
-	log.Info("generating summary of from the report file",
-		"reportFilePath", logFileSyncCombinedReport.Path,
-		"documentation", "https://rclone.org/commands/rclone_sync/")
-	err = OutputReportSummary(log, logFileSyncCombinedReport.Path)
-	if err != nil {
-		return merry.Errorf("failed to generate output report: %w", err)
-	}
-
-	// Get all the files in destination dir after syncing data
-	log.Info("getting ls data for destination directory after syncing data",
-		"path", backupConfig.DestDir)
-	err = RunRcloneLsJson(log, logFileDestFilesAfterSync.File, rcloneBinary,
-		rcloneConfig, backupConfig.DestDir)
-	if err != nil {
-		return merry.Errorf("failed to sync source and destination: %w", err)
-	}
-
-	return nil
-}
-
-func RunRcloneLsJson(log logr.Logger, commandOutputLogFile *os.File,
-	rcloneBinary, rcloneConfig, dir string,
-) error {
-	cmd := exec.Command(rcloneBinary, "lsjson",
-		"--config", rcloneConfig,
-		"-R", dir)
-	log.Info("running command", "cmd", cmd.String())
-
-	if _, err := os.Stat(dir); errors.Is(err, os.ErrNotExist) {
-		return merry.Errorf("directory does not exist: %w", err)
-	}
-
-	cmd.Stdout = commandOutputLogFile
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return merry.Errorf("failed to run command: %s: %w", string(output), err)
-	}
-	log.Info("finished running command", "cmd", cmd.String())
-
-	return nil
-}
-
-func RunRcloneSync(log logr.Logger,
-	logFileSyncPath, logFileSyncCombinedReportPath string,
-	rcloneBinary, rcloneConfig, sourceDir, destDir string,
-	extraSyncArgs []string,
-) error {
-	cmdArgs := []string{
-		"sync",
-		sourceDir, destDir,
-		"--config", rcloneConfig,
-		"--use-json-log",
-		"--log-level", "DEBUG",
-		"--log-file", logFileSyncPath,
-		"--combined", logFileSyncCombinedReportPath,
-		"--check-first",
-		"--metadata",
-	}
-	cmdArgs = append(cmdArgs, extraSyncArgs...)
-	cmd := exec.Command(rcloneBinary, cmdArgs...)
-
-	log.Info("running command", "cmd", cmd.String())
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return merry.Errorf("failed to run command: %s: %w", string(output), err)
-	}
-	log.Info("finished running command", "cmd", cmd.String())
-
-	return nil
-}
-
-func main() {
-	// TODO implement better logging?
-	// TODO context based stuff
-
-	scriptArgs, err := parseArgs()
-	if err != nil {
-		panic(fmt.Errorf("failed to parse args: %w", err))
-	}
-
-	err = validateArgs(scriptArgs)
-	if err != nil {
-		panic(fmt.Errorf("failed to validate args: %w", err))
-	}
-
+func (scriptArgs *ScriptArgs) run() {
 	logBundleDir := path.Join(scriptArgs.LogBundleBaseDir,
 		io.GetDateTimePrefixForFile()+"-log-bundle")
 	if _, err := os.Stat(logBundleDir); errors.Is(err, os.ErrNotExist) {
@@ -321,7 +123,7 @@ func main() {
 
 	log.Info("log bundle directory", "filePath", logBundleDir)
 
-	config, err := parseConfigFile(scriptArgs.Config)
+	config, err := config.ParseConfigFile(scriptArgs.Config)
 	if err != nil {
 		log.Error(err, "failed to parse config")
 		os.Exit(1)
@@ -339,15 +141,15 @@ func main() {
 	}
 
 	log.Info("getting rclone version")
-	err = RunRcloneVersion(log, scriptArgs.RcloneBinary, scriptArgs.RcloneConfig)
+	err = rclone.RunVersion(log, scriptArgs.RcloneBinary, scriptArgs.RcloneConfig)
 	if err != nil {
 		log.Error(err, "failed to get rclone version")
 		os.Exit(1)
 	}
 
-	for _, backupConfig := range config.Items {
+	for _, backupConfig := range config.BackupConfigItem {
 		log.Info("processing backup item", "config", backupConfig)
-		err := SyncSourceAndDestination(log, logBundleDir,
+		err := rclone.SyncSourceAndDestination(log, logBundleDir,
 			scriptArgs.RcloneBinary, scriptArgs.RcloneConfig, extraSyncArgs,
 			backupConfig)
 		if err != nil {
@@ -359,38 +161,20 @@ func main() {
 	}
 }
 
-func OutputReportSummary(log logr.Logger, reportFilePath string) error {
-	file, err := os.Open(reportFilePath)
+func main() {
+	// TODO context based stuff
+	// todo create rclone object and associate functions to it?
+
+	scriptArgs := &ScriptArgs{}
+	err := scriptArgs.parseArgs()
 	if err != nil {
-		return merry.Errorf("failed to open report file")
+		panic(fmt.Errorf("failed to parse args: %w", err))
 	}
-	defer file.Close()
 
-	fileScanner := bufio.NewScanner(file)
-
-	fileScanner.Split(bufio.ScanLines)
-
-	for fileScanner.Scan() {
-		line := fileScanner.Text()
-		if !strings.HasPrefix(line, "=") {
-			log.Info("non-equal file from report", "line", line)
-		}
-	}
-	return nil
-}
-
-func parseConfigFile(filePath string) (*Config, error) {
-	file, err := os.Open(filePath)
+	err = scriptArgs.verifyExpectedFilesExist()
 	if err != nil {
-		return nil, merry.Errorf("failed to open config file")
-	}
-	defer file.Close()
-
-	config := &Config{}
-	jsonParser := json.NewDecoder(file)
-	if err = jsonParser.Decode(config); err != nil {
-		return nil, merry.Errorf("failed to parse config file")
+		panic(fmt.Errorf("failed to validate args: %w", err))
 	}
 
-	return config, nil
+	scriptArgs.run()
 }
